refactor: move pokemon domain
